requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
tiktoken
langchain-text-splitters
gunicorn
